refactor : 기존 PR 에서 네이밍 변경
