+++ src/main/java/ddd/caffeine/ratrip/module/place/feign/PlaceFeignService.java
@@ -7,7 +7,7 @@ import ddd.caffeine.ratrip.module.place.feign.kakao.KakaoFeignClient;
 import ddd.caffeine.ratrip.module.place.feign.kakao.model.PlaceKakaoModel;
 import ddd.caffeine.ratrip.module.place.feign.naver.NaverFeignClient;
 import ddd.caffeine.ratrip.module.place.feign.naver.model.ImageNaverModel;
-import ddd.caffeine.ratrip.module.place.service.dto.CallPlaceSearchApiDto;
+import ddd.caffeine.ratrip.module.place.model.ThirdPartySearchOption;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -42,14 +42,14 @@ public class PlaceFeignService {
 	 * @Todo : 반경 몇 km 까지 할 것인지 프론트 개발자님들과 상의.
 	 */
 	public PlaceKakaoModel readPlacesByKeywordAndCoordinate(
-		CallPlaceSearchApiDto request) {
+		ThirdPartySearchOption option) {
 
 		final String KAKAO_REQUEST_HEADER = "KakaoAK " + KAKAO_API_KEY;
 		final int PLACE_RADIUS = 5000;
 
 		return kakaoFeignClient.readPlacesByKeywordInRadius(
-			KAKAO_REQUEST_HEADER, request.getKeyword(), request.getLatitude(), request.getLongitude(), PLACE_RADIUS,
-			request.getPage());
+			KAKAO_REQUEST_HEADER, option.getKeyword(), option.getLatitude(), option.getLongitude(), PLACE_RADIUS,
+			option.getPage());
 	}
 
 	/**
