+++ src/main/java/ddd/caffeine/ratrip/module/place/model/ThirdPartySearchOption.java
@@ -1,0 +1,16 @@
+package ddd.caffeine.ratrip.module.place.model;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ThirdPartySearchOption {
+	private String keyword;
+
+	private String latitude;
+
+	private String longitude;
+
+	private int page;
+}
