+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/dto/popular/PopularPlaceResponseDto.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.module.place.presentation.dto;
+package ddd.caffeine.ratrip.module.place.presentation.dto.popular;
 
 import java.util.ArrayList;
 import java.util.List;
